refactor: extract shared error message utility (DRY)
- Create error-utils.ts with extractErrorMessage function
- Import shared function in agent-queue.ts instead of local definition
- Import shared function in agent-process.ts to replace inline logic
- Addresses CodeRabbit/Gemini review feedback about DRY violation
