+++ apps/frontend/src/main/agent/agent-queue.ts
@@ -10,6 +10,7 @@ import type { IdeationConfig, Idea } from '../../shared/types';
 import { detectRateLimit, createSDKRateLimitInfo, getProfileEnv } from '../rate-limit-detector';
 import { getAPIProfileEnv } from '../services/profile';
 import { getOAuthModeClearVars } from './env-utils';
+import { extractErrorMessage } from './error-utils';
 import { debugLog, debugError } from '../../shared/utils/debug-logger';
 import { stripAnsiCodes } from '../../shared/utils/ansi-sanitizer';
 import { parsePythonCommand } from '../python-detector';
@@ -31,40 +32,6 @@ const STATUS_MESSAGE_MAX_LENGTH = 200;
 function formatStatusMessage(log: string): string {
   if (!log) return '';
   return stripAnsiCodes(log.trim()).split('\n')[0].substring(0, STATUS_MESSAGE_MAX_LENGTH);
-}
-
-/**
- * Extract a meaningful error message from collected stderr output.
- * Looks for error patterns and returns a summary suitable for display.
- *
- * @param stderrOutput - Collected stderr output from the process
- * @param exitCode - Process exit code
- * @returns Formatted error message
- */
-function extractErrorMessage(stderrOutput: string, exitCode: number | null): string {
-  let errorMessage = `Process exited with code ${exitCode}`;
-
-  if (stderrOutput.trim()) {
-    const stderrLines = stderrOutput.trim().split('\n').filter(line => line.trim());
-    // Look for error-like patterns
-    const errorPatterns = [
-      /error[:\s]/i, /exception/i, /failed/i, /invalid/i,
-      /unauthorized/i, /forbidden/i, /timeout/i, /traceback/i
-    ];
-    const errorLines = stderrLines.filter(line =>
-      errorPatterns.some(pattern => pattern.test(line))
-    );
-    // Use error lines if found, otherwise last few lines
-    const relevantLines = errorLines.length > 0
-      ? errorLines.slice(-3)
-      : stderrLines.slice(-3);
-    if (relevantLines.length > 0) {
-      const summary = relevantLines.join('\n').trim();
-      errorMessage = summary.length > 500 ? summary.substring(0, 500) + '...' : summary;
-    }
-  }
-
-  return errorMessage;
 }
 
 /**
